projects/file_enrichment/pyproject.toml
[tool.poetry]
name = "file_enrichment"
version = "0.1.0"
description = "The file enrichment workflow pipeline"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
dapr = "^1.15.0"
dapr-ext-fastapi = "^1.15.0"
dapr-ext-workflow = "^1.15.0"
pydantic = "^2.10.4"
python-magic = "^0.4.27"
binaryornot = "^0.4.4"
py7zr = "^0.22.0"
rarfile = "^4.2"
minio = "^7.2.14"
requests = "^2.32.3"
lief = "^0.16.2"
olefile = "^0.47"
yara-x = "^1.1.0"
pypdf = "^5.4.0"
plyara = "^2.2.7"
common = { path = "../../libs/common", develop = true }
chromium = { path = "../../libs/chromium", develop = true }
file_enrichment_modules = { path = "../../libs/file_enrichment_modules", develop = true }
structlog = "^25.1.0"
colorlog = "^6.9.0"
psycopg = {extras = ["pool"], version = "^3.2.9"}
opentelemetry-instrumentation-fastapi = "^0.51b0"
opentelemetry-api = "^1.30.0"
opentelemetry-sdk = "^1.30.0"
opentelemetry-instrumentation = "^0.51b0"
opentelemetry-exporter-zipkin-json = "^1.30.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.30.0"
rigging = "^2.2.4"
pyarrow = "^19.0.1"
msoffcrypto-tool = "^5.4.2"
oletools = "^0.60.2"
pypykatz = "^0.6.11"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
